+++ packages/snap-account-service/src/index.ts
@@ -18,6 +18,7 @@ export type {
   SnapAccountServiceGetSnapsAction,
   SnapAccountServiceHandleKeyringSnapMessageAction,
   SnapAccountServiceResolveAccountAddressAction,
+  SnapAccountServiceSetSelectedAccountsAction,
 } from './SnapAccountService-method-action-types';
 export { SnapPlatformWatcher } from './SnapPlatformWatcher';
 export type { SnapPlatformWatcherConfig } from './SnapPlatformWatcher';